+++ src/input.rs
@@ -1,34 +1,18 @@
-//! Saves an intermediate representation of the input.
+//! Parses [`DeriveInput`] into something more useful.
 
 use proc_macro2::Span;
-use syn::{DeriveInput, Fields, Generics, Ident, Result};
+use syn::{DeriveInput, Fields, Generics, Result};
 
-use crate::{Data, Default, DeriveWhere, Error, ItemAttr, Trait, VariantAttr};
+use crate::{Data, Default, DeriveWhere, Error, Item, ItemAttr, Trait, VariantAttr};
 
 /// Parsed input.
 pub struct Input<'a> {
     /// `derive_where` attributes on the item.
     pub derive_wheres: Vec<DeriveWhere>,
-    /// Item [`Span`].
-    pub span: Span,
     /// Generics necessary to define for an `impl`.
     pub generics: &'a Generics,
     /// Fields or variants of this item.
     pub item: Item<'a>,
-}
-
-/// Fields or variants of an item.
-#[allow(clippy::large_enum_variant)]
-pub enum Item<'a> {
-    /// Enum.
-    Enum {
-        /// [`struct@Ident`] of this enum.
-        ident: &'a Ident,
-        /// Variants of this enum.
-        variants: Vec<Data<'a>>,
-    },
-    /// Struct, tuple struct or union.
-    Item(Data<'a>),
 }
 
 impl<'a> Input<'a> {
@@ -151,59 +135,8 @@ impl<'a> Input<'a> {
 
         Ok(Self {
             derive_wheres,
-            span,
             generics,
             item,
         })
-    }
-}
-
-impl Item<'_> {
-    /// Return [`struct@Ident`] of this [`Item`].
-    pub fn ident(&self) -> &Ident {
-        match self {
-            Item::Item(data) => data.ident,
-            Item::Enum { ident, .. } => ident,
-        }
-    }
-
-    /// Returns `true` if any field is skipped.
-    fn any_skip(&self) -> bool {
-        match self {
-            Item::Item(data) => data.any_skip(),
-            Item::Enum { variants, .. } => variants.iter().any(|data| data.any_skip()),
-        }
-    }
-
-    /// Returns `true` if any field uses `default`.
-    // MSRV: `matches!` was added in 1.42.0.
-    #[allow(clippy::match_like_matches_macro)]
-    fn any_default(&self, derive_wheres: &[DeriveWhere]) -> bool {
-        (match self {
-            Item::Enum { .. } => true,
-            _ => false,
-        }) && derive_wheres.iter().any(|derive_where| {
-            derive_where
-                .traits
-                .iter()
-                .any(|trait_| **trait_ == Trait::Default)
-        })
-    }
-
-    /// Returns `true` if any field uses `Zeroize(fqs)`.
-    #[cfg(feature = "zeroize")]
-    fn any_fqs(&self) -> bool {
-        use crate::util::Either;
-
-        match self {
-            Item::Item(data) => match data.fields() {
-                Either::Left(fields) => fields.fields.iter().any(|field| field.attr.zeroize_fqs.0),
-                Either::Right(_) => false,
-            },
-            Item::Enum { variants, .. } => variants.iter().any(|data| match data.fields() {
-                Either::Left(fields) => fields.fields.iter().any(|field| field.attr.zeroize_fqs.0),
-                Either::Right(_) => false,
-            }),
-        }
     }
 }
